Create UserFrame.java
Shall continue to work on this, can be used to develop the JPanel that users will interact with. Once the MVC has been set up and explain shall look more into the development of it. For now shall develop the structure and set up potential interactions for users and the system (e.g. change traffic light speed)